services/api-poller/requirements.txt
# EUMETSAT API Client
eumdac==3.1.1

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Scheduling
apscheduler==3.10.4

# Redis
redis==5.0.1
# Celery for task sending
celery==5.3.4

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# HTTP Client
requests==2.31.0

# Configuration
pyyaml==6.0.1

# Monitoring
prometheus-client==0.19.0

# Logging
python-json-logger==2.0.7

# Utilities
python-dateutil==2.8.2
pytz==2023.3
